Enable FLOAT for Hive connector
* Make buckets, partitions and type utils aware of FLOAT type
* Support Parquet, ORC, Textfile, RCFILE and RC BINARY file types
* Update TupleDomainParquetPredicate such that FLOAT predicates
continue to be pushed down
* Update TupleDomainOrcPredicate such that FLOAT predicates
continue to be pushed down. For ORC, both float and double
statistics are stored in DoubleStatistics
(https://cwiki.apache.org/confluence/display/Hive/LanguageManual+ORC)

+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveType.java
@@ -41,6 +41,7 @@ import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
 import static com.facebook.presto.spi.type.DateType.DATE;
 import static com.facebook.presto.spi.type.DecimalType.createDecimalType;
 import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
+import static com.facebook.presto.spi.type.FloatType.FLOAT;
 import static com.facebook.presto.spi.type.IntegerType.INTEGER;
 import static com.facebook.presto.spi.type.SmallintType.SMALLINT;
 import static com.facebook.presto.spi.type.TimestampType.TIMESTAMP;
@@ -251,7 +252,7 @@ public final class HiveType
             case LONG:
                 return BIGINT;
             case FLOAT:
-                return DOUBLE;
+                return FLOAT;
             case DOUBLE:
                 return DOUBLE;
             case STRING:
